ACT-969: Elaborated documentation of JavaDelegate use in CustomServiceTasks
Added more details to the description of how to configure a JavaDelegate
which will be invoked when reaching a CustomServiceTask. Referred to
existing detailed descriptions for JavaDelegate and field injection
elsewhere in the userguide and linked to the new extended section from
one of the earlier sections that briefly address runtime properties.