ACT 4025 Sequence Flow bug fix
This fix is for ACT 4025 execution listeners on sequence flows not
being imported into Explorer from a Designer file.

Added a check to SequenceFlowJsonConverter class to add Execution
Listeners to the properties node on XML —> JSON conversion inside
convertToJson method. Was unable to consistently reproduce the issue in
which execution listeners were not converted to XML on export.